requirements.txt
Django==6.0
fpdf2==2.7.0
pdfrw==0.4
qrcode==7.4.2
python-multipart==0.0.6
gunicorn==21.2.0
whitenoise==6.6.0
